1374 - Update source repos list
Previous repo listings in carma-platform.repos used SSH to clone the repos,
which lead to build errors because there was no SSH key to use. The repo
URLs have been replaced with the HTTPS versions so they can be cloned inside
the container. All listed repos are public, so it does not cause any privacy
issues.

carma-platform.repos
* replace all SSH URLs with HTTPS ones so the repos can be cloned inside the
  Docker container.